adjust rsc exports
