+++ _internal/src/index.ts
@@ -10,6 +10,7 @@ export * from './utils/env'
 export { SWRGlobalState } from './utils/global-state'
 export { stableHash } from './utils/hash'
 export * from './utils/helper'
+export * from './utils/shared'
 export { mergeConfigs } from './utils/merge-config'
 export { internalMutate } from './utils/mutate'
 export { normalize } from './utils/normalize-args'
