+++ _internal/src/utils/use-swr-config.ts
@@ -1,7 +1,7 @@
 import { useContext } from 'react'
 import { defaultConfig } from './config'
 import { SWRConfigContext } from './config-context'
-import { mergeObjects } from './helper'
+import { mergeObjects } from './shared'
 import type { FullConfiguration } from '../types'
 
 export const useSWRConfig = (): FullConfiguration => {
